assignment1/requirements.txt
spacy
pandas
streamlit
pillow
pytesseract
docx2txt